requirements.txt
flask>=2.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.20
pyyaml>=6.0
python-dotenv>=1.0.0
